requirements.txt
PyYAML==6.0.2
pydantic==2.9.2
colorama==0.4.6
langchain-core
langchain-ollama
langchain-openai
langchain-community
langchain-chroma
chromadb==0.5.3
python-dotenv==1.0.1
Flask==3.0.3
Flask-SocketIO==5.3.6
